carbon_service/requirements.txt
fastapi
uvicorn
redis
httpx
python-dotenv 